+++ tests/test_httpx_sync.py
@@ -1,3 +1,4 @@
+import re
 from typing import Optional
 
 import pytest
@@ -7,7 +8,7 @@ from httpx import content_streams
 from pytest_httpx import httpx_mock, HTTPXMock
 
 
-def test_httpx_mock_without_response(httpx_mock: HTTPXMock):
+def test_without_response(httpx_mock: HTTPXMock):
     with pytest.raises(Exception) as exception_info:
         with httpx.Client() as client:
             client.get("http://test_url")
@@ -17,7 +18,7 @@ def test_httpx_mock_without_response(httpx_mock: HTTPXMock):
     )
 
 
-def test_httpx_mock_default_response(httpx_mock: HTTPXMock):
+def test_default_response(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url")
 
     with httpx.Client() as client:
@@ -28,7 +29,7 @@ def test_httpx_mock_default_response(httpx_mock: HTTPXMock):
     assert response.http_version == "HTTP/1.1"
 
 
-def test_httpx_mock_with_one_response(httpx_mock: HTTPXMock):
+def test_with_one_response(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", data=b"test content")
 
     with httpx.Client() as client:
@@ -39,7 +40,7 @@ def test_httpx_mock_with_one_response(httpx_mock: HTTPXMock):
         assert response.content == b"test content"
 
 
-def test_httpx_mock_with_many_responses(httpx_mock: HTTPXMock):
+def test_with_many_responses(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", data=b"test content 1")
     httpx_mock.add_response("http://test_url", data=b"test content 2")
 
@@ -54,7 +55,7 @@ def test_httpx_mock_with_many_responses(httpx_mock: HTTPXMock):
         assert response.content == b"test content 2"
 
 
-def test_httpx_mock_with_many_responses_methods(httpx_mock: HTTPXMock):
+def test_with_many_responses_methods(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", method="GET", data=b"test content 1")
     httpx_mock.add_response("http://test_url", method="POST", data=b"test content 2")
     httpx_mock.add_response("http://test_url", method="PUT", data=b"test content 3")
@@ -82,7 +83,7 @@ def test_httpx_mock_with_many_responses_methods(httpx_mock: HTTPXMock):
         assert response.content == b"test content 4"
 
 
-def test_httpx_mock_with_many_responses_status_codes(httpx_mock: HTTPXMock):
+def test_with_many_responses_status_codes(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         "http://test_url", method="GET", data=b"test content 1", status_code=200
     )
@@ -128,7 +129,7 @@ def test_httpx_mock_with_many_responses_status_codes(httpx_mock: HTTPXMock):
         assert response.status_code == 303
 
 
-def test_httpx_mock_with_many_responses_urls_str(httpx_mock: HTTPXMock):
+def test_with_many_responses_urls_str(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         "http://test_url?param1=test", method="GET", data=b"test content 1"
     )
@@ -170,7 +171,80 @@ def test_httpx_mock_with_many_responses_urls_str(httpx_mock: HTTPXMock):
         assert response.content == b"test content 4"
 
 
-def test_httpx_mock_with_many_responses_urls_instances(httpx_mock: HTTPXMock):
+def test_response_with_pattern_in_url(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(re.compile(".*test.*"))
+    httpx_mock.add_response("http://unmatched", data=b"test content")
+
+    with httpx.Client() as client:
+        response = client.get("http://unmatched")
+        assert response.content == b"test content"
+
+        response = client.get("http://test_url")
+        assert response.content == b""
+
+
+def test_request_with_pattern_in_url(httpx_mock: HTTPXMock):
+    httpx_mock.add_response("http://test_url")
+    httpx_mock.add_response("http://unmatched")
+
+    with httpx.Client() as client:
+        client.get("http://unmatched")
+        client.get("http://test_url", headers={"X-Test": "1"})
+
+    assert httpx_mock.get_request(re.compile(".*test.*")).headers["x-test"] == "1"
+
+
+def test_requests_with_pattern_in_url(httpx_mock: HTTPXMock):
+    httpx_mock.add_response("http://test_url")
+    httpx_mock.add_response("http://tests_url")
+    httpx_mock.add_response("http://unmatched")
+
+    with httpx.Client() as client:
+        client.get("http://tests_url", headers={"X-Test": "1"})
+        client.get("http://unmatched", headers={"X-Test": "2"})
+        client.get("http://test_url")
+
+    requests = httpx_mock.get_requests(re.compile(".*test.*"))
+    assert len(requests) == 2
+    assert requests[0].headers["x-test"] == "1"
+    assert "x-test" not in requests[1].headers
+
+
+def test_callback_with_pattern_in_url(httpx_mock: HTTPXMock):
+    def custom_response(
+        request: httpx.Request, timeout: Optional[httpx.Timeout], *args, **kwargs
+    ) -> httpx.Response:
+        return httpx.Response(
+            status_code=200,
+            http_version="HTTP/1.1",
+            headers=[],
+            stream=content_streams.JSONStream({"url": str(request.url)}),
+            request=request,
+        )
+
+    def custom_response2(
+        request: httpx.Request, timeout: Optional[httpx.Timeout], *args, **kwargs
+    ) -> httpx.Response:
+        return httpx.Response(
+            status_code=200,
+            http_version="HTTP/2.0",
+            headers=[],
+            stream=content_streams.JSONStream({"url": str(request.url)}),
+            request=request,
+        )
+
+    httpx_mock.add_callback(custom_response, re.compile(".*test.*"))
+    httpx_mock.add_callback(custom_response2, "http://unmatched")
+
+    with httpx.Client() as client:
+        response = client.get("http://unmatched")
+        assert response.http_version == "HTTP/2.0"
+
+        response = client.get("http://test_url")
+        assert response.http_version == "HTTP/1.1"
+
+
+def test_with_many_responses_urls_instances(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         httpx.URL("http://test_url", params={"param1": "test"}),
         method="GET",
@@ -222,7 +296,7 @@ def test_httpx_mock_with_many_responses_urls_instances(httpx_mock: HTTPXMock):
         assert response.content == b"test content 4"
 
 
-def test_httpx_mock_with_http_version_2(httpx_mock: HTTPXMock):
+def test_with_http_version_2(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         "http://test_url", http_version="HTTP/2", data=b"test content 1"
     )
@@ -233,7 +307,7 @@ def test_httpx_mock_with_http_version_2(httpx_mock: HTTPXMock):
         assert response.http_version == "HTTP/2"
 
 
-def test_httpx_mock_with_headers(httpx_mock: HTTPXMock):
+def test_with_headers(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         "http://test_url", data=b"test content 1", headers={"X-Test": "Test value"}
     )
@@ -244,7 +318,7 @@ def test_httpx_mock_with_headers(httpx_mock: HTTPXMock):
         assert response.headers == httpx.Headers({"x-test": "Test value"})
 
 
-def test_httpx_mock_multipart_body(httpx_mock: HTTPXMock):
+def test_multipart_body(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", data={"key1": "value1"})
     httpx_mock.add_response(
         "http://test_url",
@@ -285,7 +359,7 @@ content of file 1\r
         )
 
 
-def test_httpx_mock_requests_retrieval(httpx_mock: HTTPXMock):
+def test_requests_retrieval(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", method="GET", data=b"test content 1")
     httpx_mock.add_response("http://test_url", method="POST", data=b"test content 2")
     httpx_mock.add_response("http://test_url", method="PUT", data=b"test content 3")
@@ -324,7 +398,7 @@ def test_httpx_mock_requests_retrieval(httpx_mock: HTTPXMock):
     )
 
 
-def test_httpx_mock_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXMock):
+def test_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url")
 
     with httpx.Client() as client:
@@ -337,7 +411,7 @@ def test_httpx_mock_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXM
     assert requests[1].headers["x-test"] == "test header 2"
 
 
-def test_httpx_mock_requests_json_body(httpx_mock: HTTPXMock):
+def test_requests_json_body(httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         "http://test_url", json=["list content 1", "list content 2"]
     )
@@ -359,7 +433,7 @@ def test_httpx_mock_requests_json_body(httpx_mock: HTTPXMock):
 
 def test_callback_raising_exception(httpx_mock: HTTPXMock):
     def raise_timeout(
-        request: httpx.Request, timeout: Optional[httpx.Timeout]
+        request: httpx.Request, timeout: Optional[httpx.Timeout], *args, **kwargs
     ) -> httpx.Response:
         raise httpx.exceptions.TimeoutException()
 
@@ -372,7 +446,7 @@ def test_callback_raising_exception(httpx_mock: HTTPXMock):
 
 def test_callback_returning_response(httpx_mock: HTTPXMock):
     def custom_response(
-        request: httpx.Request, timeout: Optional[httpx.Timeout]
+        request: httpx.Request, timeout: Optional[httpx.Timeout], *args, **kwargs
     ) -> httpx.Response:
         return httpx.Response(
             status_code=200,
@@ -393,7 +467,7 @@ def test_callback_returning_response(httpx_mock: HTTPXMock):
     raises=AssertionError,
     reason="Single request cannot be returned if there is more than one matching.",
 )
-def test_httpx_mock_request_retrieval_with_more_than_one(httpx_mock: HTTPXMock):
+def test_request_retrieval_with_more_than_one(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url")
 
     with httpx.Client() as client: